chore: move schema in a separated file
